pyproject.toml
[project]
name = "ea-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
    "openai>=1.52.2",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langchain-qdrant>=0.2.1",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "qdrant-client>=1.15.1",
    "rapidfuzz>=3.14.1",
    "rdflib>=7.1.4",
    "tenacity>=9.1.2",
    "typer>=0.12.3",
    "pytest>=8.4.2",
]

[project.scripts]
ea-rag = "ea_rag.cli:main"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = [
    "tests",
]
